test: add disabledTime test
